+++ CHANGELOG.md
@@ -56,6 +56,7 @@ Compatibility:
 * Implement `Coverage.running?` method (@andrykonchin).
 * Fix arguments implicit type conversion for `Enumerable#zip` and `Array#zip` (#2788, @andrykonchin).
 * Fix `Array#unshift` to not depend on `Array#[]=` and allow overriding `#[]=` in a subclass (#2772, @andrykonchin).
+* Fix syntactic check for `void value expression` (#2821, @eregon).
 
 Performance: